pyproject.toml
[project]
name = "quenching-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torchmodel",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "numpy>=2.2.3",
    "tensordict>=0.7.1",
    "pyarrow>=19.0.1",
    "scikit-learn>=1.6.1",
    "matplotlib>=3.10.0",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
torchmodel = { workspace = true }
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'Darwin'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'Linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform == 'Darwin'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'Linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true


[tool.uv.workspace]
members = ["torchmodel"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
